+++ flake.nix
@@ -205,7 +205,7 @@
 
     # Desktop Shell
     noctalia = {
-      url = "github:noctalia-dev/noctalia/v5.0.0-beta.8";
+      url = "github:noctalia-dev/noctalia/v5.0.0-beta.10";
       inputs.nixpkgs.follows = "nixpkgs";
     };